FORMULARIO L2 VALVULAS/requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
cryptography
